+++ src/h3lib/lib/baseCells.c
@@ -18,6 +18,7 @@
  */
 
 #include "baseCells.h"
+#include "h3Index.h"
 
 /** @struct BaseCellOrient
  *  @brief base cell at a given ijk and required rotations into its system
@@ -885,4 +886,26 @@ Direction _getBaseCellDirection(int originBaseCell, int neighboringBaseCell) {
         }
     }
     return INVALID_DIGIT;
+}
+
+/**
+ * res0IndexCount returns the number of resolution 0 indexes
+ *
+ * @return int count of resolution 0 indexes
+ */
+int H3_EXPORT(res0IndexCount)() { return NUM_BASE_CELLS; }
+
+/**
+ * getRes0Indexes generates all base cells storing them into the provided
+ * memory pointer. Buffer must be of size NUM_BASE_CELLS * sizeof(H3Index).
+ *
+ * @param out H3Index* the memory to store the resulting base cells in
+ */
+void H3_EXPORT(getRes0Indexes)(H3Index* out) {
+    for (int bc = 0; bc < NUM_BASE_CELLS; bc++) {
+        H3Index baseCell = H3_INIT;
+        H3_SET_MODE(baseCell, H3_HEXAGON_MODE);
+        H3_SET_BASE_CELL(baseCell, bc);
+        out[bc] = baseCell;
+    }
 }