+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The public API of this library consists of the functions declared in file
 [h3api.h](./src/h3lib/include/h3api.h).
 
 ## [Unreleased]
+### Added
+- Include defined constants for current library version
 
 ## [3.3.0] - 2018-12-25
 ### Added